move training and inference script to bin/
